добавил update метод

+++ routes/api_v1.php
@@ -21,7 +21,7 @@ Route::prefix('book')->group(function() {
     Route::get("/raw", [BookController::class, 'all'])->name('get_all_books');
     Route::get("/{book_id}", [BookController::class, 'get'])->name('get_one_book')->where(['book_id' => '[0-9]+']);
     Route::post("/", [BookController::class, 'create'])->name('create_book');
-//    Route::put("/{book_id}", [\App\Http\Controllers\Api\V1\BookController::class, 'update'])->name('update_book')->where(['book_id' => '[0-9]+']);
+    Route::put("/{book_id}", [BookController::class, 'update'])->name('update_book')->where(['book_id' => '[0-9]+']);
     Route::delete("/{book_id}", [BookController::class, 'delete'])->name('delete_one_book')->where(['book_id' => '[0-9]+']);
     Route::post("/{book_id}/author", [BookController::class, 'addAuthor'])->name('add_author')->where(['book_id' => '[0-9]+']);
     Route::get("/{book_id}/authors", [BookController::class, 'authors'])->name('get_book_authors')->where(['book_id' => '[0-9]+']);
@@ -33,11 +33,11 @@ Route::prefix('book')->group(function() {
 Route::prefix('author')->group(function() {
     Route::get("/", [AuthorController::class, 'paginate'])->name('get_all_authors_paginate');
     Route::get("/raw", [AuthorController::class, 'all'])->name('get_all_authors');
-    Route::get("/{author_id}", [AuthorController::class, 'get'])->name('get_one_author');
+    Route::get("/{author_id}", [AuthorController::class, 'get'])->name('get_one_author')->where(['author_id' => '[0-9]+']);
     Route::post("/", [AuthorController::class, 'create'])->name('create_author');
-    Route::delete("/{author_id}", [AuthorController::class, 'delete'])->name('delete_one_author');
-    //Route::put("/{author_id}", [AuthorController::class, 'update'])->name('update_author');
-    Route::get("/{author_id}/books", [AuthorController::class, 'books'])->name('authors_books');
+    Route::put("/{author_id}", [AuthorController::class, 'update'])->name('update_author')->where(['author_id' => '[0-9]+']);
+    Route::delete("/{author_id}", [AuthorController::class, 'delete'])->name('delete_one_author')->where(['author_id' => '[0-9]+']);
+    Route::get("/{author_id}/books", [AuthorController::class, 'books'])->name('authors_books')->where(['author_id' => '[0-9]+']);
 });
 
 Route::fallback(function () { return response('Bad Request', 400);});
